fix: allow g:template_autoload to be set to boolean value
Previous test was whether variable was zero or non-zero. This did not
allow for it to be set to v:true or v:false. Changing the test to
evaluate whether true or not allows for boolean values, which is
compatible with nvim lua plugin files setting the variable to lua true
or false.